Basic Collision Implemented
Gave buildings their own layer for collisions and set them so they collide only with water and other buildings - note that they do not collide with the ground, as that check is done by the raycast and not the individual building.  I also fixed the raycast function to match this (before it was sending the layermask to the max distance parameter).